refactor(engine): extract chat templating from inference.py (#454)
#3 maintainability: move chat-template application and message
normalization (~714 lines) into engine/chat_templating.py — tool
injection, native-tool-format hint, thinking resolution, text + VLM
apply_chat_template wrappers, segmented tokenization, and the tool-message
role rewrites. inference.py re-exports them so call sites, chat/session.py,
and tests are unchanged. Drops the now-unused json import.
